+++ event-store-management/event-store-management-core/src/main/java/uk/gov/justice/services/eventstore/management/catchup/process/EventCatchupRunner.java
@@ -36,7 +36,8 @@ public class EventCatchupRunner {
                 clock.now()
         ));
 
-        subscriptionCatchupDefinitions.forEach(catchupFor -> eventCatchupByComponentRunner.runEventCatchupForComponent(
+        subscriptionCatchupDefinitions
+                .forEach(catchupFor -> eventCatchupByComponentRunner.runEventCatchupForComponent(
                 catchupFor,
                 commandId,
                 catchupCommand));