+++ CHANGELOG.md
@@ -5,11 +5,12 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 
 ## [Unreleased]
 
+## [2.4.6] - 2020-01-21
+### Added
+- Catchup for multiple components now run in order of component and subscription priority
+- Added event source name to catchup logger output
 ### Fixed
 - Fixed catchup error where catchup was marked as complete after all subscriptions rather than all components
-
-### Added
-- Added event source name to catchup logger output
 
 ## [2.4.5] - 2020-01-06
 ### Removed